Physically moved last java8 tests to core module
