--- context-propagation-java8/src/test/resources/META-INF/services/nl.talsmasoftware.context.ContextManager
@@ -1,1 +1,0 @@
-nl.talsmasoftware.context.DummyContextManager
