--- context-propagation-java8/src/test/resources/logback-test.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration debug="true">
-
-    <contextListener class="ch.qos.logback.classic.jul.LevelChangePropagator">
-        <resetJUL>true</resetJUL>
-    </contextListener>
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level - %logger{36} - %method{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <logger name="nl.talsmasoftware" level="DEBUG" />
-
-    <root level="WARN">
-        <appender-ref ref="STDOUT"/>
-    </root>
-
-</configuration>
