+++ context-propagation/src/test/java/nl/talsmasoftware/context/functions/BinaryOperatorWithContextTest.java
@@ -24,26 +24,17 @@ import org.junit.Test;
 
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.BiFunction;
-import java.util.function.Function;
-import java.util.function.Supplier;
+import java.util.function.*;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.hasToString;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.sameInstance;
+import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.fail;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 /**
  * @author Sjoerd Talsma
  */
-public class BiFunctionWithContextTest {
+public class BinaryOperatorWithContextTest {
 
     private ContextSnapshot snapshot;
     private Context<Void> context;
@@ -68,14 +59,14 @@ public class BiFunctionWithContextTest {
 
     @Test
     public void testApply() {
-        new BiFunctionWithContext<>(snapshot, (a, b) -> b).apply("input1", "input2");
+        new BinaryOperatorWithContext<>(snapshot, (a, b) -> b).apply("input1", "input2");
         verify(snapshot).reactivate();
     }
 
     @Test
     public void testApplyWithoutSnapshot() {
         try {
-            new BiFunctionWithContext<>(null, (a, b) -> b);
+            new BinaryOperatorWithContext<>(null, (a, b) -> b);
             fail("Exception expected");
         } catch (RuntimeException expected) {
             assertThat(expected, hasToString(containsString("No context snapshot provided")));
@@ -85,7 +76,7 @@ public class BiFunctionWithContextTest {
     @Test
     public void testApplyWithoutSnapshotSupplier() {
         try {
-            new BiFunctionWithContext<>((Supplier<ContextSnapshot>) null, (a, b) -> b, context -> {
+            new BinaryOperatorWithContext<>((Supplier<ContextSnapshot>) null, (a, b) -> b, context -> {
             });
             fail("Exception expected");
         } catch (RuntimeException expected) {
@@ -98,7 +89,7 @@ public class BiFunctionWithContextTest {
         final ContextSnapshot[] snapshotHolder = new ContextSnapshot[1];
         DummyContextManager.setCurrentValue("Old value");
 
-        Thread t = new Thread(() -> new BiFunctionWithContext<>(snapshot,
+        Thread t = new Thread(() -> new BinaryOperatorWithContext<>(snapshot,
                 (input1, input2) -> {
                     DummyContextManager.setCurrentValue("New value");
                     return input2;
@@ -122,7 +113,8 @@ public class BiFunctionWithContextTest {
         final RuntimeException expectedException = new RuntimeException("Whoops!");
 
         try {
-            new BiFunctionWithContext<>(snapshot, throwing(expectedException)).apply("input1", "input2");
+            new BinaryOperatorWithContext<String>(() -> snapshot, throwing(expectedException), null) {
+            }.apply("input1", "input2");
             fail("Exception expected");
         } catch (RuntimeException rte) {
             assertThat(rte, is(sameInstance(expectedException)));
@@ -135,7 +127,7 @@ public class BiFunctionWithContextTest {
     @Test
     public void testAndThenNull() {
         try {
-            new BiFunctionWithContext<>(snapshot, (a, b) -> b).andThen(null);
+            new BinaryOperatorWithContext<>(snapshot, (a, b) -> b).andThen(null);
             fail("Exception expected");
         } catch (NullPointerException expected) {
             assertThat(expected, hasToString(containsString("after function <null>")));
@@ -145,12 +137,12 @@ public class BiFunctionWithContextTest {
     @Test
     public void testAndThen_singleContextSwitch() {
         when(snapshot.reactivate()).thenReturn(context);
-        Function<Integer, Integer> after = i -> i + 100;
-        BiFunction<Integer, Integer, Integer> function = (a, b) -> a * 10 + b * 5;
+        UnaryOperator<Integer> after = i -> i + 100;
+        BinaryOperator<Integer> function = (a, b) -> a * 10 + b * 5;
         AtomicInteger consumed = new AtomicInteger(0);
 
         BiFunction<Integer, Integer, Integer> composed =
-                new BiFunctionWithContext<>(snapshot, function, snapshot -> consumed.incrementAndGet())
+                new BinaryOperatorWithContext<>(snapshot, function, snapshot -> consumed.incrementAndGet())
                         .andThen(after);
 
         assertThat(composed.apply(2, 3), is(20 + 15 + 100));
@@ -169,7 +161,7 @@ public class BiFunctionWithContextTest {
         }
     }
 
-    private static <IN1, IN2, OUT> BiFunction<IN1, IN2, OUT> throwing(RuntimeException rte) {
+    private static <T> BinaryOperator<T> throwing(RuntimeException rte) {
         return (input1, input2) -> {
             throw rte;
         };
