+++ context-propagation/src/test/java/nl/talsmasoftware/context/futures/CompletedFutureTest.java
