+++ context-propagation/src/test/java/nl/talsmasoftware/context/futures/ContextSnapshotHolderTest.java
@@ -19,9 +19,7 @@ import nl.talsmasoftware.context.ContextSnapshot;
 import org.junit.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.hamcrest.Matchers.sameInstance;
+import static org.hamcrest.Matchers.*;
 import static org.mockito.Mockito.mock;
 
 public class ContextSnapshotHolderTest {
